move deprecated structure sections into mixins
