+++ nanome/api/structure/chain.py
@@ -1,16 +1,13 @@
 from nanome._internal.structure.models import _Chain
 from nanome.util import Logs
 from . import Base
+from ._deprecated import ChainDeprecated
 
 
-class Chain(_Chain, Base):
+class Chain(_Chain, ChainDeprecated, Base):
     """
     | Represents a Chain. Contains residues. Molecules contains chains.
     """
-
-    def __init__(self):
-        super(Chain, self).__init__()
-        self._molecular = Chain.Molecular(self)
 
     def add_residue(self, residue):
         """
@@ -95,25 +92,5 @@ class Chain(_Chain, Base):
         if type(value) is not str:
             value = str(value)
         self._name = value
-    # endregion
-
-    # region deprecated
-    @property
-    @Logs.deprecated()
-    def molecular(self):
-        return self._molecular
-
-    class Molecular(object):
-        def __init__(self, parent):
-            self.parent = parent
-
-        @property
-        def name(self):
-            return self.parent.name
-
-        @name.setter
-        def name(self, value):
-            self.parent.name = value
-
     # endregion
 _Chain._create = Chain
